Allow verification of untrusted peer certs
Current use-case:
The service fastly-config-backend is used to test changes to the GCDN
for regressions. This includes routing traffic to the correct Pantheon
region for the site, so it needs to be deployed to all our regions.
Although it isn't a production service, it runs on production kubernetes
clusters in order to have a physical presence in regions like europe. On
those clusters, the sandbox CA bundle is not available.

The Fastly services that connect to fastly-config-backend are test
services. As test services, they obtain their certificates and other
secrets from sandbox vault. Of course, this results in mTLS handshake
failures with fastly-config-backend.

Since this is only a test service used for fabricating interesting
responses, and is not in any way a part of the product, risk from an MITM
attack against it is low.

This allows the configuration of the test fastly services to obtain all
their information from only one instance of vault, significantly
simplifying the GCDN deployment machinery.